requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
django-redis==6.0.0
requests==2.32.4
drf-yasg==1.21.10